bot-service/requirements.txt
requests
python-binance
pandas
numpy
python-dotenv